Fix -Wsign-compare warnings (#883)
* Fix -Wsign-compare warnings

Currently, we see lots of warnings when building with `-DBACKEND=cuda`.
This attempts to fix -Wsign-compare warnings.

* Fix warnings in tests

+++ examples/cpp/tag-bilstm/train_tag-bilstm.cc
@@ -101,7 +101,7 @@ struct RNNLanguageModel {
           // Find best tag according to the distribution
           double best = -9e99;
           int besti = -1;
-          for (int i = 0; i < dist.size(); ++i) {
+          for (int i = 0; i < static_cast<int>(dist.size()); ++i) {
             if (dist[i] > best) { best = dist[i]; besti = i; }
           }
           if (tags[t] == besti) (*cor)++;
